requirements.txt
torch
transformers
PyMuPDF
python-docx
streamlit
